Update src/utils.jl
Add type for mdot_in

+++ src/utils.jl
@@ -186,7 +186,7 @@ function build_rho_rc(eps_m::T, eps_g::T, eps_x::T, rho_m::T, rho_g::T, rho_x::T
     return [rho, drho_dP, drho_dT, drho_deps_g, rc, drc_dP, drc_dT]
 end
 
-function build_mdot_in(fluxing::Bool, rho_m0::Number, log_vfr::Number, P_0::Number, T_in::Number)
+function build_mdot_in(fluxing::Bool, rho_m0::Number, log_vfr::Number, P_0::Number, T_in::Number)::Float64
     if ~fluxing
         range_vfr = 10^log_vfr   # volume flow rate (km3/yr)  
         mdot_in   = rho_m0*range_vfr*1e9/(3600*24*365)
